+++ src/cpu/o3/SConscript
@@ -41,7 +41,13 @@ if env['CONF']['TARGET_ISA'] != 'null':
     Source('cpu.cc')
     Source('decode.cc')
     Source('dyn_inst.cc')
-    Source('fetch.cc')
+    Source('fetch/fetch.cc')
+    Source('fetch/fetch_main.cc')
+    Source('fetch/fetch_cache.cc')
+    Source('fetch/fetch_thread.cc')
+    Source('fetch/fetch_pipeline.cc')
+    Source('fetch/fetch_branch.cc')
+    Source('fetch/fetch_dual.cc')
     Source('free_list.cc')
     Source('fu_pool.cc')
     Source('iew.cc')
