+++ src/cpu/o3/fetch/fetch.cc
@@ -1,0 +1,543 @@
+/*
+ * Copyright (c) 2010-2014 ARM Limited
+ * Copyright (c) 2012-2013 AMD
+ * All rights reserved.
+ *
+ * The license below extends only to copyright in the software and shall
+ * not be construed as granting a license to any other intellectual
+ * property including but not limited to intellectual property relating
+ * to a hardware implementation of the functionality of the software
+ * licensed hereunder.  You may use the software subject to the license
+ * terms below provided that you ensure that this notice is replicated
+ * unmodified and in its entirety in all distributions of the software,
+ * modified or unmodified, in source code or in binary form.
+ *
+ * Copyright (c) 2004-2006 The Regents of The University of Michigan
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are
+ * met: redistributions of source code must retain the above copyright
+ * notice, this list of conditions and the following disclaimer;
+ * redistributions in binary form must reproduce the above copyright
+ * notice, this list of conditions and the following disclaimer in the
+ * documentation and/or other materials provided with the distribution;
+ * neither the name of the copyright holders nor the names of its
+ * contributors may be used to endorse or promote products derived from
+ * this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+#include "cpu/o3/fetch/fetch.hh"
+
+#include <algorithm>
+#include <cstring>
+#include <list>
+#include <map>
+#include <queue>
+#include <set>
+#include <sstream>
+
+#include "arch/generic/tlb.hh"
+#include "arch/riscv/decoder.hh"
+#include "base/debug_helper.hh"
+#include "base/random.hh"
+#include "base/types.hh"
+#include "config/the_isa.hh"
+#include "cpu/base.hh"
+#include "cpu/exetrace.hh"
+#include "cpu/nop_static_inst.hh"
+#include "cpu/o3/cpu.hh"
+#include "cpu/o3/dyn_inst.hh"
+#include "cpu/o3/limits.hh"
+#include "debug/Activity.hh"
+#include "debug/Counters.hh"
+#include "debug/DecoupleBPProbe.hh"
+#include "debug/Drain.hh"
+#include "debug/Fetch.hh"
+#include "debug/FetchFault.hh"
+#include "debug/FetchVerbose.hh"
+#include "debug/O3CPU.hh"
+#include "debug/O3PipeView.hh"
+#include "mem/packet.hh"
+#include "params/BaseO3CPU.hh"
+#include "sim/byteswap.hh"
+#include "sim/core.hh"
+#include "sim/eventq.hh"
+#include "sim/full_system.hh"
+#include "sim/system.hh"
+
+namespace gem5
+{
+
+namespace o3
+{
+
+Fetch::Fetch(CPU *_cpu, const BaseO3CPUParams &params)
+    : fetchPolicy(params.smtFetchPolicy),
+      cpu(_cpu),
+      branchPred(nullptr),
+      decodeToFetchDelay(params.decodeToFetchDelay),
+      renameToFetchDelay(params.renameToFetchDelay),
+      iewToFetchDelay(params.iewToFetchDelay),
+      commitToFetchDelay(params.commitToFetchDelay),
+      fetchWidth(params.fetchWidth),
+      decodeWidth(params.decodeWidth),
+      retryPkt(),
+      retryTid(InvalidThreadID),
+      cacheBlkSize(cpu->cacheLineSize()),
+      enable2Fetch(params.enable2Fetch),
+      fetchBufferSize(params.fetchBufferSize),
+      fetchQueueSize(params.fetchQueueSize),
+      numThreads(params.numThreads),
+      numFetchingThreads(params.smtNumFetchingThreads),
+      icachePort(this, _cpu),
+      finishTranslationEvent(this), fetchStats(_cpu, this)
+{
+    if (numThreads > MaxThreads)
+        fatal("numThreads (%d) is larger than compiled limit (%d),\n"
+              "\tincrease MaxThreads in src/cpu/o3/limits.hh\n",
+              numThreads, static_cast<int>(MaxThreads));
+    if (fetchWidth > MaxWidth)
+        fatal("fetchWidth (%d) is larger than compiled limit (%d),\n"
+             "\tincrease MaxWidth in src/cpu/o3/limits.hh\n",
+             fetchWidth, static_cast<int>(MaxWidth));
+
+    for (int i = 0; i < MaxThreads; i++) {
+        setThreadStatus(i, Idle);
+        decoder[i] = nullptr;
+        pc[i].reset(params.isa[0]->newPCState());
+        macroop[i] = nullptr;
+        delayedCommit[i] = false;
+        stalls[i] = {false, false};
+        lastIcacheStall[i] = 0;
+    }
+
+    branchPred = params.branchPred;
+
+    if (isStreamPred()) {
+        dbsp = dynamic_cast<branch_prediction::stream_pred::DecoupledStreamBPU*>(branchPred);
+        dbpftb = nullptr;
+        dbpbtb = nullptr;
+        assert(dbsp);
+        usedUpFetchTargets = true;
+    } else if (isFTBPred()) {
+        dbsp = nullptr;
+        dbpftb = dynamic_cast<branch_prediction::ftb_pred::DecoupledBPUWithFTB*>(branchPred);
+        dbpbtb = nullptr;
+        assert(dbpftb);
+        usedUpFetchTargets = true;
+        dbpftb->setCpu(_cpu);
+    } else if (isBTBPred()) {
+        dbsp = nullptr;
+        dbpftb = nullptr;
+        dbpbtb = dynamic_cast<branch_prediction::btb_pred::DecoupledBPUWithBTB*>(branchPred);
+        assert(dbpbtb);
+        usedUpFetchTargets = true;
+        dbpbtb->setCpu(_cpu);
+    }
+
+    assert(params.decoder.size());
+    for (ThreadID tid = 0; tid < numThreads; tid++) {
+        decoder[tid] = params.decoder[tid];
+        // Set the size and allocate data for each fetch buffer instance (dual buffers)
+        for (unsigned ftqIndex = 0; ftqIndex < 2; ++ftqIndex) {
+            fetchBuffer[tid][ftqIndex].size = fetchBufferSize;
+            fetchBuffer[tid][ftqIndex].data = new uint8_t[fetchBufferSize];
+        }
+    }
+
+    // Get the size of an instruction.
+    // stallReason size should be the same as decodeWidth,renameWidth,dispWidth
+    stallReason.resize(decodeWidth, StallReason::NoStall);
+}
+
+std::string Fetch::name() const { return cpu->name() + ".fetch"; }
+
+void
+Fetch::regProbePoints()
+{
+    ppFetch = new ProbePointArg<DynInstPtr>(cpu->getProbeManager(), "Fetch");
+    ppFetchRequestSent = new ProbePointArg<RequestPtr>(cpu->getProbeManager(),
+                                                       "FetchRequest");
+
+}
+
+Fetch::FetchStatGroup::FetchStatGroup(CPU *cpu, Fetch *fetch)
+    : statistics::Group(cpu, "fetch"),
+    ADD_STAT(icacheStallCycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch is stalled on an Icache miss"),
+    ADD_STAT(insts, statistics::units::Count::get(),
+             "Number of instructions fetch has processed"),
+    ADD_STAT(branches, statistics::units::Count::get(),
+             "Number of branches that fetch encountered"),
+    ADD_STAT(predictedBranches, statistics::units::Count::get(),
+             "Number of branches that fetch has predicted taken"),
+    ADD_STAT(cycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch has run and was not squashing or "
+             "blocked"),
+    ADD_STAT(squashCycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch has spent squashing"),
+    ADD_STAT(tlbCycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch has spent waiting for tlb"),
+    ADD_STAT(idleCycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch was idle"),
+    ADD_STAT(blockedCycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch has spent blocked"),
+    ADD_STAT(miscStallCycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch has spent waiting on interrupts, or bad "
+             "addresses, or out of MSHRs"),
+    ADD_STAT(pendingDrainCycles, statistics::units::Cycle::get(),
+             "Number of cycles fetch has spent waiting on pipes to drain"),
+    ADD_STAT(noActiveThreadStallCycles, statistics::units::Cycle::get(),
+             "Number of stall cycles due to no active thread to fetch from"),
+    ADD_STAT(pendingTrapStallCycles, statistics::units::Cycle::get(),
+             "Number of stall cycles due to pending traps"),
+    ADD_STAT(pendingQuiesceStallCycles, statistics::units::Cycle::get(),
+             "Number of stall cycles due to pending quiesce instructions"),
+    ADD_STAT(icacheWaitRetryStallCycles, statistics::units::Cycle::get(),
+             "Number of stall cycles due to full MSHR"),
+    ADD_STAT(cacheLines, statistics::units::Count::get(),
+             "Number of cache lines fetched"),
+    ADD_STAT(icacheSquashes, statistics::units::Count::get(),
+             "Number of outstanding Icache misses that were squashed"),
+    ADD_STAT(tlbSquashes, statistics::units::Count::get(),
+             "Number of outstanding ITLB misses that were squashed"),
+    ADD_STAT(nisnDist, statistics::units::Count::get(),
+             "Number of instructions fetched each cycle (Total)"),
+    ADD_STAT(idleRate, statistics::units::Ratio::get(),
+             "Ratio of cycles fetch was idle",
+             idleCycles / cpu->baseStats.numCycles),
+    ADD_STAT(branchRate, statistics::units::Ratio::get(),
+             "Number of branch fetches per cycle",
+             branches / cpu->baseStats.numCycles),
+    ADD_STAT(rate, statistics::units::Rate<
+                    statistics::units::Count, statistics::units::Cycle>::get(),
+             "Number of inst fetches per cycle",
+             insts / cpu->baseStats.numCycles),
+    ADD_STAT(fetchStatusDist, statistics::units::Count::get(),
+             "Distribution of fetch status"),
+    ADD_STAT(decodeStalls, statistics::units::Count::get(),
+             "Number of decode stalls"),
+    ADD_STAT(decodeStallRate, statistics::units::Rate<
+                    statistics::units::Count, statistics::units::Cycle>::get(),
+             "Number of decode stalls per cycle",
+             decodeStalls / cpu->baseStats.numCycles),
+    ADD_STAT(fetchBubbles, statistics::units::Count::get(),
+             "Unutilized issue-pipeline slots while there is no backend-stall"),
+    ADD_STAT(fetchBubbles_max, statistics::units::Count::get(),
+             "Cycles that fetch 0 instruction while there is no backend-stall"),
+    ADD_STAT(frontendBound, statistics::units::Rate<
+                    statistics::units::Count, statistics::units::Cycle>::get(),
+             "Frontend Bound",
+             fetchBubbles / (cpu->baseStats.numCycles * fetch->decodeWidth)),
+    ADD_STAT(frontendLatencyBound, statistics::units::Rate<
+                    statistics::units::Count, statistics::units::Cycle>::get(),
+             "Frontend Latency Bound",
+             fetchBubbles_max / cpu->baseStats.numCycles),
+    ADD_STAT(frontendBandwidthBound, statistics::units::Rate<
+                    statistics::units::Count, statistics::units::Cycle>::get(),
+             "Frontend Bandwidth Bound",
+             frontendBound - frontendLatencyBound),
+    ADD_STAT(twoFetchRequests, statistics::units::Count::get(),
+             "Number of 2 fetch requests"),
+    ADD_STAT(twoFetchSuccess, statistics::units::Count::get(),
+             "Number of 2 fetch requests that are successful"),
+    ADD_STAT(twoFetchFailedDueToBankConflict, statistics::units::Count::get(),
+             "Number of 2 fetch requests that are failed due to bank conflicts"),
+    ADD_STAT(twoFetchFailedDueToFallThrough, statistics::units::Count::get(),
+             "Number of 2 fetch requests that are failed due to fall through"),
+    ADD_STAT(twoFetchFailedDueToInvalidFTQPCs, statistics::units::Count::get(),
+             "Number of 2 fetch requests that are failed due to invalid FTQ PCs"),
+    ADD_STAT(twoFetchFailedDueToFetchRangesSpanMultiplePages, statistics::units::Count::get(),
+             "Number of 2 fetch requests that are failed due to fetch ranges span multiple pages"),
+    ADD_STAT(twoFetchFailedDueToNoNewFTQEntry, statistics::units::Count::get(),
+             "Number of 2 fetch requests that are failed due to no new FTQ entry"),
+    ADD_STAT(twoFetchFailedDueToCacheBlocked, statistics::units::Count::get(),
+             "Number of 2 fetch requests that are failed due to cache blocked")
+{
+        icacheStallCycles
+            .prereq(icacheStallCycles);
+        insts
+            .prereq(insts);
+        branches
+            .prereq(branches);
+        predictedBranches
+            .prereq(predictedBranches);
+        cycles
+            .prereq(cycles);
+        squashCycles
+            .prereq(squashCycles);
+        tlbCycles
+            .prereq(tlbCycles);
+        idleCycles
+            .prereq(idleCycles);
+        blockedCycles
+            .prereq(blockedCycles);
+        cacheLines
+            .prereq(cacheLines);
+        miscStallCycles
+            .prereq(miscStallCycles);
+        pendingDrainCycles
+            .prereq(pendingDrainCycles);
+        noActiveThreadStallCycles
+            .prereq(noActiveThreadStallCycles);
+        pendingTrapStallCycles
+            .prereq(pendingTrapStallCycles);
+        pendingQuiesceStallCycles
+            .prereq(pendingQuiesceStallCycles);
+        icacheWaitRetryStallCycles
+            .prereq(icacheWaitRetryStallCycles);
+        icacheSquashes
+            .prereq(icacheSquashes);
+        tlbSquashes
+            .prereq(tlbSquashes);
+        nisnDist
+            .init(/* base value */ 0,
+              /* last value */ fetch->fetchWidth,
+              /* bucket size */ 1)
+            .flags(statistics::pdf);
+        idleRate
+            .prereq(idleRate);
+        branchRate
+            .flags(statistics::total);
+        rate
+            .flags(statistics::total);
+        fetchStatusDist
+            .init(NumFetchStatus)
+            .flags(statistics::pdf | statistics::total);
+
+        for (int i = 0; i < NumFetchStatus; i++) {
+            fetchStatusDist.subname(i, fetch->fetchStatusStr[static_cast<Fetch::ThreadStatus>(i)]);
+        }
+        decodeStalls
+            .prereq(decodeStalls);
+        decodeStallRate
+            .flags(statistics::total);
+        fetchBubbles
+            .prereq(fetchBubbles);
+        fetchBubbles_max
+            .prereq(fetchBubbles_max);
+        frontendBound
+            .flags(statistics::total);
+        frontendLatencyBound
+            .flags(statistics::total);
+        frontendBandwidthBound
+            .flags(statistics::total);
+}
+
+void
+Fetch::setTimeBuffer(TimeBuffer<TimeStruct> *time_buffer)
+{
+    timeBuffer = time_buffer;
+
+    // Create wires to get information from proper places in time buffer.
+    fromDecode = timeBuffer->getWire(-decodeToFetchDelay);
+    fromRename = timeBuffer->getWire(-renameToFetchDelay);
+    fromIEW = timeBuffer->getWire(-iewToFetchDelay);
+    fromCommit = timeBuffer->getWire(-commitToFetchDelay);
+}
+
+void
+Fetch::setActiveThreads(std::list<ThreadID> *at_ptr)
+{
+    activeThreads = at_ptr;
+}
+
+void
+Fetch::setFetchQueue(TimeBuffer<FetchStruct> *ftb_ptr)
+{
+    // Create wire to write information to proper place in fetch time buf.
+    toDecode = ftb_ptr->getWire(0);
+
+    // initialize to toDecode stall vector
+    toDecode->fetchStallReason = stallReason;
+}
+
+void
+Fetch::startupStage()
+{
+    assert(priorityList.empty());
+    resetStage();
+
+    // Fetch needs to start fetching instructions at the very beginning,
+    // so it must start up in active state.
+    switchToActive();
+}
+
+void
+Fetch::clearStates(ThreadID tid)
+{
+    setThreadStatus(tid, Running);
+    set(pc[tid], cpu->pcState(tid));
+    macroop[tid] = NULL;
+    delayedCommit[tid] = false;
+    // Reset all FTQ cache requests and fetch buffers
+    for (unsigned ftqIndex = 0; ftqIndex < 2; ++ftqIndex) {
+        cacheReq[tid][ftqIndex].reset();
+        fetchBuffer[tid][ftqIndex].reset();
+    }
+    stalls[tid].decode = false;
+    stalls[tid].drain = false;
+    fetchQueue[tid].clear();
+
+    // TODO not sure what to do with priorityList for now
+    // priorityList.push_back(tid);
+}
+
+void
+Fetch::resetStage()
+{
+    numInst = 0;
+    interruptPending = false;
+    cacheBlocked = false;
+
+    priorityList.clear();
+
+    // Setup PC and nextPC with initial state.
+    for (ThreadID tid = 0; tid < numThreads; ++tid) {
+        setThreadStatus(tid, Running);
+        set(pc[tid], cpu->pcState(tid));
+        macroop[tid] = NULL;
+
+        delayedCommit[tid] = false;
+        // Reset all FTQ cache requests and fetch buffers
+        for (unsigned ftqIndex = 0; ftqIndex < 2; ++ftqIndex) {
+            cacheReq[tid][ftqIndex].reset();
+            fetchBuffer[tid][ftqIndex].reset();
+        }
+
+        stalls[tid].decode = false;
+        stalls[tid].drain = false;
+
+        fetchQueue[tid].clear();
+
+        priorityList.push_back(tid);
+    }
+
+    wroteToTimeBuffer = false;
+    _status = Inactive;
+
+    // Initialize usedUpFetchTargets to force getting initial FTQ entry
+    usedUpFetchTargets = true;
+
+    DPRINTF(Fetch, "resetStage: set usedUpFetchTargets=true for initial FTQ setup\n");
+
+    if (isStreamPred()) {
+        dbsp->resetPC(pc[0]->instAddr());
+    } else if (isFTBPred()) {
+        dbpftb->resetPC(pc[0]->instAddr());
+    } else if (isBTBPred()) {
+        dbpbtb->resetPC(pc[0]->instAddr());
+    }
+}
+
+void
+Fetch::drainResume()
+{
+    for (ThreadID i = 0; i < numThreads; ++i) {
+        stalls[i].decode = false;
+        stalls[i].drain = false;
+    }
+}
+
+void
+Fetch::drainSanityCheck() const
+{
+    assert(isDrained());
+    assert(retryPkt.size() == 0);
+    assert(retryTid == InvalidThreadID);
+    assert(!cacheBlocked);
+    assert(!interruptPending);
+
+    for (ThreadID i = 0; i < numThreads; ++i) {
+        for (unsigned ftqIndex = 0; ftqIndex < 2; ++ftqIndex) {
+            assert(cacheReq[i][ftqIndex].packets.empty());
+        }
+        assert(fetchStatus[i] == Idle || stalls[i].drain);
+    }
+
+    branchPred->drainSanityCheck();
+}
+
+bool
+Fetch::isDrained() const
+{
+    /* Make sure that threads are either idle of that the commit stage
+     * has signaled that draining has completed by setting the drain
+     * stall flag. This effectively forces the pipeline to be disabled
+     * until the whole system is drained (simulation may continue to
+     * drain other components).
+     */
+    for (ThreadID i = 0; i < numThreads; ++i) {
+        // Verify fetch queues are drained
+        if (!fetchQueue[i].empty())
+            return false;
+
+        // Return false if not idle or drain stalled
+        if (fetchStatus[i] != Idle) {
+            if (fetchStatus[i] == Blocked && stalls[i].drain)
+                continue;
+            else
+                return false;
+        }
+    }
+
+    /* The pipeline might start up again in the middle of the drain
+     * cycle if the finish translation event is scheduled, so make
+     * sure that's not the case.
+     */
+    return !finishTranslationEvent.scheduled();
+}
+
+void
+Fetch::takeOverFrom()
+{
+    assert(cpu->getInstPort().isConnected());
+    resetStage();
+
+}
+
+void
+Fetch::drainStall(ThreadID tid)
+{
+    assert(cpu->isDraining());
+    assert(!stalls[tid].drain);
+    DPRINTF(Drain, "%i: Thread drained.\n", tid);
+    stalls[tid].drain = true;
+}
+
+void
+Fetch::wakeFromQuiesce()
+{
+    DPRINTF(Fetch, "Waking up from quiesce\n");
+    // Hopefully this is safe
+    // @todo: Allow other threads to wake from quiesce.
+    setThreadStatus(0, Running);
+}
+
+void
+Fetch::flushFetchBuffer()
+{
+    for (ThreadID i = 0; i < numThreads; ++i) {
+        for (unsigned ftqIndex = 0; ftqIndex < 2; ++ftqIndex) {
+            FetchBuffer& bufferRef = getFetchBuffer(i, ftqIndex);
+            bufferRef.valid = false;
+        }
+    }
+}
+
+
+} // namespace o3
+} // namespace gem5
