+++ src/cpu/o3/fetch/fetch_thread.cc
@@ -1,0 +1,290 @@
+#include <algorithm>
+#include <list>
+#include <map>
+
+#include "base/types.hh"
+#include "cpu/o3/cpu.hh"
+#include "cpu/o3/fetch/fetch.hh"
+#include "cpu/o3/limits.hh"
+#include "debug/Activity.hh"
+#include "debug/Fetch.hh"
+#include "mem/packet.hh"
+#include "params/BaseO3CPU.hh"
+#include "sim/byteswap.hh"
+#include "sim/core.hh"
+#include "sim/eventq.hh"
+#include "sim/full_system.hh"
+#include "sim/system.hh"
+
+namespace gem5
+{
+
+namespace o3
+{
+
+ThreadID
+Fetch::getFetchingThread()
+{
+    if (numThreads > 1) {
+        switch (fetchPolicy) {
+          case SMTFetchPolicy::RoundRobin:
+            return roundRobin();
+          case SMTFetchPolicy::IQCount:
+            return iqCount();
+          case SMTFetchPolicy::LSQCount:
+            return lsqCount();
+          case SMTFetchPolicy::Branch:
+            return branchCount();
+          default:
+            return InvalidThreadID;
+        }
+    } else {
+        std::list<ThreadID>::iterator thread = activeThreads->begin();
+        if (thread == activeThreads->end()) {
+            return InvalidThreadID;
+        }
+
+        ThreadID tid = *thread;
+
+        if (canFetchInstructions(tid) || fetchStatus[tid] == Idle) {
+            return tid;
+        } else {
+            return InvalidThreadID;
+        }
+    }
+}
+
+ThreadID
+Fetch::selectFetchThread()
+{
+    ThreadID tid = getFetchingThread();
+
+    assert(!cpu->switchedOut());
+
+    if (tid == InvalidThreadID) {
+        // Breaks looping condition in tick()
+        threadFetched = numFetchingThreads;
+
+        if (numThreads == 1) {
+            profileStall(0);
+        }
+        return InvalidThreadID;
+    }
+
+    return tid;
+}
+
+ThreadID
+Fetch::roundRobin()
+{
+    std::list<ThreadID>::iterator pri_iter = priorityList.begin();
+    std::list<ThreadID>::iterator end      = priorityList.end();
+
+    ThreadID high_pri;
+
+    while (pri_iter != end) {
+        high_pri = *pri_iter;
+
+        assert(high_pri <= numThreads);
+
+        if (canFetchInstructions(high_pri) || fetchStatus[high_pri] == Idle) {
+
+            priorityList.erase(pri_iter);
+            priorityList.push_back(high_pri);
+
+            return high_pri;
+        }
+
+        pri_iter++;
+    }
+
+    return InvalidThreadID;
+}
+
+ThreadID
+Fetch::iqCount()
+{
+    //sorted from lowest->highest
+    std::priority_queue<unsigned, std::vector<unsigned>,
+                        std::greater<unsigned> > PQ;
+    std::map<unsigned, ThreadID> threadMap;
+
+    std::list<ThreadID>::iterator threads = activeThreads->begin();
+    std::list<ThreadID>::iterator end = activeThreads->end();
+
+    while (threads != end) {
+        ThreadID tid = *threads++;
+        unsigned iqCount = cpu->getIQInsts();
+
+        //we can potentially get tid collisions if two threads
+        //have the same iqCount, but this should be rare.
+        PQ.push(iqCount);
+        threadMap[iqCount] = tid;
+    }
+
+    while (!PQ.empty()) {
+        ThreadID high_pri = threadMap[PQ.top()];
+
+        if (canFetchInstructions(high_pri) || fetchStatus[high_pri] == Idle)
+            return high_pri;
+        else
+            PQ.pop();
+
+    }
+
+    return InvalidThreadID;
+}
+
+ThreadID
+Fetch::lsqCount()
+{
+    //sorted from lowest->highest
+    std::priority_queue<unsigned, std::vector<unsigned>,
+                        std::greater<unsigned> > PQ;
+    std::map<unsigned, ThreadID> threadMap;
+
+    std::list<ThreadID>::iterator threads = activeThreads->begin();
+    std::list<ThreadID>::iterator end = activeThreads->end();
+
+    while (threads != end) {
+        ThreadID tid = *threads++;
+        unsigned ldstqCount = fromIEW->iewInfo[tid].ldstqCount;
+
+        //we can potentially get tid collisions if two threads
+        //have the same iqCount, but this should be rare.
+        PQ.push(ldstqCount);
+        threadMap[ldstqCount] = tid;
+    }
+
+    while (!PQ.empty()) {
+        ThreadID high_pri = threadMap[PQ.top()];
+
+        if (canFetchInstructions(high_pri) || fetchStatus[high_pri] == Idle)
+            return high_pri;
+        else
+            PQ.pop();
+    }
+
+    return InvalidThreadID;
+}
+
+ThreadID
+Fetch::branchCount()
+{
+    panic("Branch Count Fetch policy unimplemented\n");
+    return InvalidThreadID;
+}
+
+void
+Fetch::setThreadStatus(ThreadID tid, ThreadStatus status)
+{
+    assert(tid < MaxThreads);
+
+    ThreadStatus oldStatus = fetchStatus[tid];
+    fetchStatus[tid] = status;
+    DPRINTF(Fetch, "[tid:%d] setThreadStatus: %s -> %s\n", tid, fetchStatusStr[oldStatus], fetchStatusStr[status]);
+}
+
+bool
+Fetch::checkStall(ThreadID tid) const
+{
+    bool ret_val = false;
+
+    if (stalls[tid].drain) {
+        assert(cpu->isDraining());
+        DPRINTF(Fetch,"[tid:%i] Drain stall detected.\n",tid);
+        ret_val = true;
+    }
+
+    return ret_val;
+}
+
+void
+Fetch::switchToActive()
+{
+    if (_status == Inactive) {
+        DPRINTF(Activity, "Activating stage.\n");
+
+        cpu->activateStage(CPU::FetchIdx);
+
+        _status = Active;
+    }
+}
+
+void
+Fetch::switchToInactive()
+{
+    if (_status == Active) {
+        DPRINTF(Activity, "Deactivating stage.\n");
+
+        cpu->deactivateStage(CPU::FetchIdx);
+
+        _status = Inactive;
+    }
+}
+
+void
+Fetch::deactivateThread(ThreadID tid)
+{
+    // Update priority list
+    auto thread_it = std::find(priorityList.begin(), priorityList.end(), tid);
+    if (thread_it != priorityList.end()) {
+        priorityList.erase(thread_it);
+    }
+}
+
+void
+Fetch::profileStall(ThreadID tid)
+{
+    DPRINTF(Fetch,"There are no more threads available to fetch from.\n");
+
+    // @todo Per-thread stats
+
+    if (stalls[tid].drain) {
+        ++fetchStats.pendingDrainCycles;
+        DPRINTF(Fetch, "Fetch is waiting for a drain!\n");
+    } else if (activeThreads->empty()) {
+        ++fetchStats.noActiveThreadStallCycles;
+        DPRINTF(Fetch, "Fetch has no active thread!\n");
+    } else if (fetchStatus[tid] == Blocked) {
+        ++fetchStats.blockedCycles;
+        DPRINTF(Fetch, "[tid:%i] Fetch is blocked!\n", tid);
+    } else if (fetchStatus[tid] == Squashing) {
+        ++fetchStats.squashCycles;
+        DPRINTF(Fetch, "[tid:%i] Fetch is squashing!\n", tid);
+    } else if (getOverallCacheStatus(tid) == CacheWaitResponse) {
+        ++fetchStats.icacheStallCycles;
+        DPRINTF(Fetch, "[tid:%i] Fetch is waiting cache response!\n",
+                tid);
+    } else if (getOverallCacheStatus(tid) == TlbWait) {
+        ++fetchStats.tlbCycles;
+        DPRINTF(Fetch, "[tid:%i] Fetch is waiting ITLB walk to "
+                "finish!\n", tid);
+    } else if (fetchStatus[tid] == TrapPending) {
+        ++fetchStats.pendingTrapStallCycles;
+        DPRINTF(Fetch, "[tid:%i] Fetch is waiting for a pending trap!\n",
+                tid);
+    } else if (getOverallCacheStatus(tid) == CacheWaitRetry) {
+        ++fetchStats.icacheWaitRetryStallCycles;
+        DPRINTF(Fetch, "[tid:%i] Fetch is waiting for an I-cache retry!\n",
+                tid);
+    } else if (getOverallCacheStatus(tid) == AccessFailed) {
+            DPRINTF(Fetch, "[tid:%i] Fetch predicted non-executable address\n",
+                    tid);
+    } else {
+        DPRINTF(Fetch, "[tid:%i] Unexpected fetch stall reason "
+            "(Status: %i)\n",
+            tid, fetchStatus[tid]);
+    }
+}
+
+void
+Fetch::setAllFetchStalls(StallReason stall)
+{
+    for (int i = 0; i < stallReason.size(); i++) {
+        stallReason[i] = stall;
+    }
+}
+
+} // namespace o3
+} // namespace gem5
